add multi-tool session activity: OpenCode + source badges + session detail overlay
- loadOpenCodeActivity() reads ~/.local/share/opencode/opencode.db via
  sqlite3 CLI, scoped to current project worktree
- loadClaudeSessionActivity() now tags rows with source="claude" and
  sessionID (file path) for drill-down
- loadAgents() merges maple/claude/opencode rows, cap 15
- agentsContent() shows [cc]/[oc]/[ml] source badge per row
- Enter on Agents pane opens sessionDetailView() overlay with scrollable
  tool/prompt timeline (j/k/g/G/Esc)
- loadClaudeSessionLines() reads JSONL and formats tool_use entries
- loadOpenCodeSessionLines() queries sqlite parts for text/tool/step-start
- SOLID split: dashboard_views.go and loaders.go (new files)
